web/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-autoslug==1.9.9
djangorestframework==3.14.0
drf-nested-routers==0.93.5
idna==3.6
numpy==1.26.4
pandas==2.2.0
pgeocode==0.4.1
progressbar2==4.3.2
psycopg==3.1.18
psycopg-binary==3.1.18
python-dateutil==2.8.2
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
urllib3==2.2.0
